+++ common.moplaf.macroplanner/com.misc.common.moplaf.macroplanner.solver.emf.edit/src/com/misc/common/moplaf/macroplanner/solver/provider/LPRoutingItemProvider.java
@@ -206,7 +206,7 @@ public class LPRoutingItemProvider extends LPTimeLineItemProvider  implements II
 
 		@Override
 		public float getEventAmount(LPRoutingBucket bucket, int moment) {
-			return bucket.getPlannedSelectedSolution();
+			return (float) bucket.getPlannedSelectedSolution();
 		}
 
 		@Override
