+++ common.moplaf.macroplanner/com.misc.common.moplaf.macroplanner.solver.emf/src/com/misc/common/moplaf/macroplanner/solver/LPRoutingBucket.java
@@ -144,6 +144,6 @@ public interface LPRoutingBucket extends LPTimeBucket {
 	 * @model transient="true" changeable="false" volatile="true" derived="true"
 	 * @generated
 	 */
-	float getPlannedSelectedSolution();
+	double getPlannedSelectedSolution();
 
 } // LPRoutingBucket
